restructure models and add regiter them to admin
